requirements.txt
azure-mgmt-resource==3.0.0
haikunator
